Update for Fondant 0.7.0 (#4)
I already published Fondant 0.7.0 to `test.pypi` so we can upgrade our
examples and test them against it (which I did). The main change is the
split of the component and pipeline SDKs.

This PR also includes some improvements to the notebook which we already
did in the other repositories.